Created Views for song sharing received and sent page - Refactored the TrackView, TrackDetailsView, TrackList to be their own views and build upon each other - Currently rendering TrackList in the SongShareView, but will need to create new SharedTrackList and SharedTrack views     - This new view will contain additional info such as the sender/receiver and maybe more - Fixed padding bug in ProfileView
