+++ spoti-friends/src/profile/views/UserProfileView.swift
@@ -98,6 +98,7 @@ struct UserProfileView: View {
                     }
                     Spacer()
                 }
+                .padding(.horizontal, 20)
                 .onAppear {
                     Task {
                         // Comment out these lines for SwiftUI Previews
